+++ src/pages/RecoveryPassword/Answer/Answer.module.css
@@ -1,0 +1,17 @@
+.check {
+    font-style: normal;
+    font-weight: 600;
+    font-size: 22px;
+    line-height: 33px;
+    color: #2D2E46;
+    margin-top: 24px;
+    margin-bottom: 22px;
+}
+
+.text {
+    font-style: normal;
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 24px;
+    color: #9293a2;
+}
